latest revision of Bounce.sh script
Used to bounce the Wifi/VPN/and network adapters on Mac OS, related to a bug where the wifi still shows associated with an Access point, but won't obtain internet access.  It usually happens when it wakes up from sleep mode, and is associated with a new AP/wifi network than it was before the Macbook went to sleep.